wip: merge test files
